--- a/src/front.rs
+++ b/src/front.rs
@@ -2,9 +2,8 @@
 //! on the user side, allowing to add more tasks to the queue.
 
 use std::sync::Arc;
-use {Handle, Wait, Task};
+use {Handle, Wait, TaskBuilder};
 use back::Backend;
-use pulse::Signal;
 
 
 /// Queue front-end.
@@ -37,19 +36,11 @@ pub trait Schedule {
     /// Add a new task with selected dependencies. This doesn't interrupt any
     /// tasks in-flight. The task will actually start as soon as all 
     /// dependencies are finished.
-    fn add_task(&self, t: Box<Task+Send>, signal: Option<Signal>) -> Handle;
-
-    /// Create a new task that is a child of the parent task.
-    /// A child task will extend the lifespan of the parent
-    fn add_child_task(&self, t: Box<Task+Send>, signal: Option<Signal>) -> Handle;
+    fn add_task(&self, task: TaskBuilder) -> Handle;
 }
 
 impl Schedule for Frontend {
-    fn add_task(&self, task: Box<Task+Send>, signal: Option<Signal>) -> Handle {
-        Backend::start(self.backend.clone(), task, signal)
-    }
-
-    fn add_child_task(&self, task: Box<Task+Send>, signal: Option<Signal>) -> Handle {
-        Backend::start(self.backend.clone(), task, signal)  
+    fn add_task(&self, task: TaskBuilder) -> Handle {
+        Backend::start(self.backend.clone(), task, None)
     }
 }