This adds an idle pool to the backend. The idea is that a thread is spawned for work, once it is done it will try and see new work. This works much like a threadpool except in one important difference. This will spawn a new thread if it does not find one immediately ready to take the request.

This helps with performance:
```
/// Before
test chain_10_use_die    ... bench:    309695 ns/iter (+/- 68397)
test chain_10_wait       ... bench:    307926 ns/iter (+/- 23675)
test chain_1_000_use_die ... bench:  30693362 ns/iter (+/- 1686540)
test chain_1_000_wait    ... bench:  30737621 ns/iter (+/- 1060133)
test fibb_depth_6        ... bench:  11248402 ns/iter (+/- 3322318)
test start_die           ... bench:       104 ns/iter (+/- 0)

/// After
test chain_10_use_die    ... bench:    191749 ns/iter (+/- 9538)
test chain_10_wait       ... bench:     85817 ns/iter (+/- 8267)
test chain_1_000_use_die ... bench:   8031486 ns/iter (+/- 545938)
test chain_1_000_wait    ... bench:   8006532 ns/iter (+/- 1061809)
test fibb_depth_6        ... bench:   3751003 ns/iter (+/- 2192736)
test start_die           ... bench:       436 ns/iter (+/- 1)
```